Remove hardcoded armv7-a architecture flag
The current Makefile assumes that the target is an armv7-a compatible platform (ie i.MX6) when KERNELRELEASE is set. This will cause a build failure on architectures that do not support the armv7-a instruction set such as the NXP QoriQ, Layerscape or 64-bit i.MX range of devices, all of which support CAAM.
Note that this change only applies to the caam_keyblob driver and does not fix builds of caam_tool, which will still fail due to the hard-coded GOARCH value.